backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pydantic>=1.8.2
scrapy>=2.6.0
python-dotenv>=0.19.0
requests>=2.26.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=3.2.0
